SAMSON-285 now the verbose levels impose lesser levels. Example '-vvv' forces '-vv' and '-v' as well.
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@2186 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/parseArgs/baStd.h
@@ -54,7 +54,7 @@
 *
 * baStoi - string to integer
 */
-extern int baStoi(char* string);
+extern long long baStoi(char* string);
